refactor: eliminate unnecessary panics from production code
Replace explicit panic! macros with more idiomatic alternatives where
the panic guards invariant violations that are provably unreachable:

- build_expression_from_list: debug_assert! (callers guard empty case)
- build_balanced_boolean_expression: debug_assert! (public API returns
  Option for empty)
- Automaton::empty: .expect() on hardcoded sentinel pattern
- DEFAULT_INDEX_BUILD_POLICY: .expect() on compile-time-embedded TOML
- copyright patterns regex: .expect() on hardcoded regex strings

Restructure post_processing::test_utils to remove panic-containing
code from the public API:

- Move golden-test-only helpers (7 assert_* functions with panics,
  plus supporting utilities) into new golden_helpers module gated
  #[cfg(feature = "golden-tests")] pub mod
- Trim test_utils to only file/dir/package/scan_and_assemble helpers,
  now #[cfg(test)] mod (private, test-only)
- Update integration test imports to use golden_helpers
